Support For Choosing Mesh, Material and Texture
Adding ModelOptionsController which holds the available Meshes, Materials and Textures that can be applied to the model. It will then apply them when the corresponding button is selected.

Initial UI support for a ListView was added so we can display the different options that will be available to change the model.